#2049 Move tests to reach full coverage for each class.
