+++ src/psyclone/tests/domain/common/base_driver_creator_test.py
@@ -40,6 +40,7 @@
 import pytest
 
 from psyclone.domain.common import BaseDriverCreator
+from psyclone.domain.lfric import LFRicExtractDriverCreator
 from psyclone.domain.lfric.transformations import LFRicExtractTrans
 from psyclone.psyir.nodes import (Assignment, Literal, Routine,
                                   StructureReference)
@@ -232,3 +233,37 @@ def test_base_driver_structure_accesses():
         _ = psy.gen
     assert ("The provided PSyIR should not have StructureReferences, "
             "but found: x_ptr_vector%method" in str(err.value))
+
+
+# ----------------------------------------------------------------------------
+def test_lfric_driver_import_modules():
+    '''Tests that adding a call detects errors as expected.
+    '''
+    program = Routine.create("routine", is_program=True)
+    _, invoke = get_invoke("8_vector_field_2.f90", "lfric",
+                           dist_mem=False, idx=0)
+
+    sched = invoke.schedule
+    # We need to lower to convert the kernels to calls
+    sched.lower_to_language_level()
+    # and add them to the extraction driver
+    program.children.extend([node.copy() for node in sched.children])
+
+    driver_creator = LFRicExtractDriverCreator()
+
+    # Initially we should only have no symbol other than the routine:
+    assert ['routine'] == [sym.name for sym in program.symbol_table.symbols]
+
+    driver_creator.import_modules(program)
+    # We should now have two more symbols:
+    all_symbols = ["routine",
+                   "testkern_coord_w0_2_mod",
+                   "testkern_coord_w0_2_code"]
+    assert (all_symbols == [sym.name for sym in program.symbol_table.symbols])
+
+    # Import twice so we test the handling of symbols that
+    # are already in the symbol table:
+    driver_creator.import_modules(program)
+
+    # The symbol table should be the same as it was before:
+    assert (all_symbols == [sym.name for sym in program.symbol_table.symbols])
